+++ macros/src/bevy.rs
@@ -1,20 +1,21 @@
-use proc_macro2::{Ident, Span, TokenStream, TokenTree};
+use super::get_key_raw;
+use proc_macro2::{Ident, Span, TokenStream, TokenTree, Literal};
 use proc_macro_error::abort;
 use quote::quote;
 use std::borrow::Cow;
 
-// pub fn get_key(tree: TokenTree) -> Option<TokenStream> {
-//     get_key_raw(tree).map(|r| match r {
-//         Ok(c) => {
-//             let l = Literal::character(c);
-//             quote! { ::winit::keyboard::Key::Character(#l) }
-//         },
-//         Err(cow) => {
-//             let i = Ident::new(&cow, Span::call_site());
-//             quote! { ::winit::keyboard::Key::Named(::winit::keyboard::NamedKey::#i) }
-//         }
-//     })
-// }
+pub fn get_key(tree: TokenTree) -> Option<TokenStream> {
+    get_key_raw(tree).map(|r| match r {
+        Ok(c) => {
+            let l = Literal::string(&c.to_string());
+            quote! { ::bevy::input::keyboard::Key::Character(#l.into()) }
+        },
+        Err(cow) => {
+            let i = Ident::new(&cow, Span::call_site());
+            quote! { ::bevy::input::keyboard::Key::#i }
+        }
+    })
+}
 
 pub fn get_pkey(tree: TokenTree) -> Option<TokenStream> {
     match tree {